Total utility is now a ratio, 50/50 trades are better if ahead

+++ src/main/java/com/MoveUtility.java
@@ -3,15 +3,15 @@ package com;
 public class MoveUtility implements Comparable<MoveUtility> {
 
     Move move;
-    int utilityDifference;  // Always positive (Black is trying to get the max)
+    double utilityDifference;  // Always positive (Black is trying to get the max)
 
-    public MoveUtility (Move move, int utilityDifference) {
+    public MoveUtility (Move move, double utilityDifference) {
         this.move = move;
         this.utilityDifference = utilityDifference;
     }
 
     @Override
     public int compareTo(MoveUtility other) {
-        return Integer.compare(utilityDifference, other.utilityDifference);
+        return Double.compare(utilityDifference, other.utilityDifference);
     }
 }
